// heroesApp/heroesApp/05-heroesApp/src/app/heroes/interfaces/heroes.interface.ts
export interface Heroe {
    id?:               string;
    superhero:        string;
    publisher:        Publisher;
    alter_ego:        string;
    first_appearance: string;
    characters:       string;
    alt_img?:         string; //http://image.com/image.png ......
}

export enum Publisher {
    DCComics = "DC Comics",
    MarvelComics = "Marvel Comics",
}